git-svn-id: svn://192.168.1.10/public/CoffeeMud@10919 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Libraries/CMGenEditor.java
@@ -8600,8 +8600,8 @@ public class CMGenEditor extends StdLibrary implements GenericEditor
 			me.setEfficiencyPct(CMath.div(prompt(mob,Math.round(me.getEfficiencyPct()*100),++showNumber,showFlag,"Efficiency % (50-150): "),100.0));
 			me.setReliabilityPct(CMath.div(prompt(mob,Math.round(me.getReliabilityPct()*100),++showNumber,showFlag,"Reliability % (0-100): "),100.0));
 			me.setItemMask(prompt(mob,me.getItemMaskStr(),++showNumber,showFlag,"Item Mask (?): "));
-			me.setMinTechLevelDiff((byte)prompt(mob,(short)me.getMinTechLevelDiff(),++showNumber,showFlag,"Min Tech Diff: "));
-			me.setMaxTechLevelDiff((byte)prompt(mob,(short)me.getMaxTechLevelDiff(),++showNumber,showFlag,"Max Tech Diff: "));
+			me.setMinTechLevelDiff((byte)prompt(mob,me.getMinTechLevelDiff(),++showNumber,showFlag,"Min Tech Diff: "));
+			me.setMaxTechLevelDiff((byte)prompt(mob,me.getMaxTechLevelDiff(),++showNumber,showFlag,"Max Tech Diff: "));
 			if(me.getMaxTechLevelDiff()<me.getMinTechLevelDiff())
 				me.setMaxTechLevelDiff(me.getMinTechLevelDiff());
 			me.setManufactureredTypesList(promptCommaList(mob,me.getManufactureredTypesList(),++showNumber,showFlag,"Manufact. Types: ",
